feat move fetchtasks to common
